Flags of type int for syscall clone

+++ src/args_trace.c
@@ -360,7 +360,7 @@ void get_args_clone(process_descriptor_t * proc, reg_s * reg, syscall_arg_u * sy
 {
   clone_arg_t arg = &(sysarg->clone);
   arg->ret = reg->ret;
-  arg->clone_flags = reg->arg[0];
+  arg->clone_flags = (int) reg->arg[0];
   arg->newsp = reg->arg[1];
   arg->parent_tid = (void *) reg->arg[2];
   arg->child_tid = (void *) reg->arg[3];
